Update documentation
It is tricky to implement realloc with only malloc and free, so remove
the plans for dima/malloc_and_free.h. However, dima/realloc_and_free.h
is still possible, so use that in README.md instead.

The environment entry point should be dima_init_env(), and not
dima_env_instance().